Improve repayment token (#113)
* rename borrowingToken -> repaymentToken

* add repayment argument

* use amounts in contract for redemption

* make helper ZERO and helper collateral getting

* test repayment with different decimals

* decimalize and make bonds different to get

* format

* merge typo

* remove repaymentRatio, rename collateralToken to backingToken, use IERC20Metadata for decimals querying, fix the tests and add some comments

+++ contracts/test/TestERC20.sol
@@ -9,7 +9,7 @@ contract TestERC20 is ERC20 {
     constructor(
         string memory name,
         string memory symbol,
-        uint96 _mintAmount,
+        uint256 _mintAmount,
         uint8 decimals_
     ) ERC20(name, symbol) {
         _mint(msg.sender, _mintAmount);
